requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
scikit-learn
flask
scipy
opencv-python-headless
